feat: add event user info queries(#109)
